Added github/gitignore as submodule
When you start a new project it is often necessary to add a custom
`.gitignore` file to your repository, depending on the content of
your project or its language. The github/gitignore repository has
some handy templates to get started. I like to keep a copy of it
in my home directory.
This submodule creates a folder called `.gitignore.d` that you can
symlink from your home directory.